[feature/#32] 회고 기록 bottomsheet 재수정 - 처음감정
