+++ app/src/main/java/com/teampome/pome/presentation/remind/RemindFirstBottomSheetFragment.kt
@@ -1,0 +1,34 @@
+package com.teampome.pome.presentation.remind
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import com.teampome.pome.databinding.FragmentRemindFirstBottomSheetBinding
+
+class RemindFirstBottomSheetFragment : BottomSheetDialogFragment() {
+    private var _binding: FragmentRemindFirstBottomSheetBinding? = null
+    private val binding get() = _binding ?: error("binding이 되지 않았습니다")
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentRemindFirstBottomSheetBinding.inflate(layoutInflater, container, false)
+
+        initHeight()
+
+        return binding.root
+    }
+
+    private fun initHeight() {
+        val bottomSheet = binding.clRemindFirstBottomsheet
+        bottomSheet.minHeight = 259
+        bottomSheet.maxHeight = 259
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
