+++ app/src/main/java/com/teampome/pome/presentation/remind/RemindFragment.kt
@@ -10,15 +10,16 @@ import com.teampome.pome.util.FriendsConsumeItemDecorator
 
 class RemindFragment : BaseFragment<FragmentRemindBinding>(R.layout.fragment_remind) {
     private lateinit var remindConsumeAdapter: RemindConsumeAdapter
-    private val bottomSheetFragment = RemindBottomSheetFragment()
+    private val bottomSheetFragment = RemindFirstBottomSheetFragment()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
         initRemindConsumeAdapter()
         initAdapterDecoration()
-        initAddRemindConsume() //서버통신할 코드
-        showBottomSheet()
+        initAddRemindConsume()
+
+        initClickFirstEmotion()
     }
 
     private fun initAdapterDecoration() {
@@ -110,18 +111,12 @@ class RemindFragment : BaseFragment<FragmentRemindBinding>(R.layout.fragment_rem
         )
     }
 
-    private fun showBottomSheet() {
-        //서버 통신을 통해 반응이 있는 경우만 보여주기
-        remindConsumeAdapter.setReactionClickListener(object :
-            RemindConsumeAdapter.ReactionClickListener {
-            override fun onClick(data: View, pos: Int) {
-                if (!bottomSheetFragment.isAdded)
-                    bottomSheetFragment.show(
-                        childFragmentManager, bottomSheetFragment.tag
-                    )
-            }
-
-        })
-
+    private fun initClickFirstEmotion() {
+        binding.ivFirstEmotion.setOnClickListener {
+            if (!bottomSheetFragment.isAdded)
+                bottomSheetFragment.show(
+                    childFragmentManager, bottomSheetFragment.tag
+                )
+        }
     }
 }
