+++ app/src/main/res/drawable/ic_emoji_mint_sad_filter_54.xml
@@ -1,0 +1,102 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="54dp"
+    android:height="54dp"
+    android:viewportWidth="54"
+    android:viewportHeight="54">
+  <path
+      android:pathData="M53.08,31.734C56.903,44.785 44.785,56.903 31.734,53.08C28.577,52.155 25.219,52.155 22.062,53.08C9.011,56.903 -3.107,44.785 0.716,31.734C1.641,28.577 1.641,25.219 0.716,22.062C-3.107,9.011 9.011,-3.107 22.062,0.716C25.219,1.641 28.577,1.641 31.734,0.716C44.785,-3.107 56.903,9.011 53.08,22.062C52.155,25.222 52.155,28.577 53.08,31.734Z">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="5.406"
+          android:startY="5.406"
+          android:endX="52.064"
+          android:endY="52.065"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFA0F2B7"/>
+        <item android:offset="0.058" android:color="#FF99EFAF"/>
+        <item android:offset="0.225" android:color="#FF8AE89D"/>
+        <item android:offset="0.346" android:color="#FF85E597"/>
+        <item android:offset="0.36" android:color="#FF83E395"/>
+        <item android:offset="0.535" android:color="#FF6ED486"/>
+        <item android:offset="0.651" android:color="#FF67CE80"/>
+        <item android:offset="1" android:color="#FF5FB772"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M36.653,0.002V1.54C41.502,1.54 46.173,3.908 49.143,7.875C52.11,11.834 53.005,16.849 51.605,21.632C50.594,25.078 50.594,28.721 51.605,32.167C53.005,36.95 52.107,41.965 49.143,45.924C46.17,49.891 41.502,52.259 36.653,52.259C35.159,52.259 33.648,52.04 32.167,51.604C30.456,51.102 28.684,50.848 26.901,50.848C25.118,50.848 23.346,51.102 21.635,51.604C20.154,52.038 18.643,52.259 17.149,52.259C12.3,52.259 7.629,49.891 4.659,45.924C1.692,41.965 0.797,36.95 2.197,32.167C3.208,28.721 3.208,25.078 2.197,21.632C0.797,16.849 1.695,11.834 4.659,7.875C7.632,3.908 12.3,1.54 17.149,1.54C18.643,1.54 20.154,1.759 21.635,2.195C23.346,2.697 25.118,2.951 26.901,2.951C28.684,2.951 30.456,2.697 32.167,2.195C33.648,1.761 35.159,1.54 36.653,1.54V0.002ZM36.653,0.002C35.049,0.002 33.4,0.23 31.734,0.719C30.154,1.181 28.526,1.411 26.898,1.411C25.27,1.411 23.642,1.181 22.062,0.719C20.397,0.232 18.747,0.002 17.144,0.002C6.181,0.002 -2.62,10.679 0.716,22.065C1.641,25.222 1.641,28.58 0.716,31.737C-2.62,43.122 6.178,53.799 17.144,53.799C18.747,53.799 20.397,53.572 22.062,53.083C23.642,52.62 25.27,52.39 26.898,52.39C28.526,52.39 30.154,52.62 31.734,53.083C33.4,53.569 35.049,53.799 36.653,53.799C47.616,53.799 56.417,43.122 53.08,31.737C52.155,28.58 52.155,25.222 53.08,22.065C56.417,10.677 47.619,0.002 36.653,0.002Z"
+      android:strokeAlpha="0.8"
+      android:fillAlpha="0.8">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="26.899"
+          android:startY="-5.984"
+          android:endX="26.899"
+          android:endY="63.06"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="0.243" android:color="#FFFDFDFD"/>
+        <item android:offset="0.386" android:color="#FFF4F8F5"/>
+        <item android:offset="0.504" android:color="#FFE6EEE8"/>
+        <item android:offset="0.608" android:color="#FFD2E0D5"/>
+        <item android:offset="0.703" android:color="#FFB8CEBC"/>
+        <item android:offset="0.79" android:color="#FF98B89E"/>
+        <item android:offset="0.873" android:color="#FF719E7B"/>
+        <item android:offset="0.95" android:color="#FF468152"/>
+        <item android:offset="1" android:color="#FF266B34"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M35.139,24.962C36.481,24.962 37.569,23.875 37.569,22.532C37.569,21.191 36.481,20.103 35.139,20.103C33.797,20.103 32.709,21.191 32.709,22.532C32.709,23.875 33.797,24.962 35.139,24.962Z"
+      android:fillColor="#1A211B"/>
+  <path
+      android:pathData="M18.661,24.962C20.003,24.962 21.091,23.875 21.091,22.532C21.091,21.191 20.003,20.103 18.661,20.103C17.319,20.103 16.231,21.191 16.231,22.532C16.231,23.875 17.319,24.962 18.661,24.962Z"
+      android:fillColor="#1A211B"/>
+  <path
+      android:pathData="M34.314,22.324C34.686,22.324 34.988,22.022 34.988,21.65C34.988,21.278 34.686,20.977 34.314,20.977C33.942,20.977 33.64,21.278 33.64,21.65C33.64,22.022 33.942,22.324 34.314,22.324Z"
+      android:strokeAlpha="0.5"
+      android:fillColor="#5FB772"
+      android:fillAlpha="0.5"/>
+  <path
+      android:pathData="M17.85,22.324C18.222,22.324 18.524,22.022 18.524,21.65C18.524,21.278 18.222,20.977 17.85,20.977C17.478,20.977 17.176,21.278 17.176,21.65C17.176,22.022 17.478,22.324 17.85,22.324Z"
+      android:strokeAlpha="0.5"
+      android:fillColor="#5FB772"
+      android:fillAlpha="0.5"/>
+  <path
+      android:pathData="M38.804,32.159C38.804,34.292 37.887,34.937 36.754,34.937C35.62,34.937 34.703,34.29 34.703,32.159C34.703,30.028 36.754,26.762 36.754,26.762C36.754,26.762 38.804,30.026 38.804,32.159Z"
+      android:strokeAlpha="0.7"
+      android:fillAlpha="0.7">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startX="36.754"
+          android:startY="26.155"
+          android:endX="36.754"
+          android:endY="38.712"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFFFFFF"/>
+        <item android:offset="0.198" android:color="#FFFCFEFD"/>
+        <item android:offset="0.341" android:color="#FFF3FCF7"/>
+        <item android:offset="0.467" android:color="#FFE3F9ED"/>
+        <item android:offset="0.583" android:color="#FFCDF4DE"/>
+        <item android:offset="0.692" android:color="#FFB0EDCB"/>
+        <item android:offset="0.795" android:color="#FF8CE5B4"/>
+        <item android:offset="0.895" android:color="#FF62DC99"/>
+        <item android:offset="0.99" android:color="#FF33D17A"/>
+        <item android:offset="1" android:color="#FF2DD076"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+  <path
+      android:pathData="M37.74,30.32C37.74,31.346 37.299,31.659 36.751,31.659C36.206,31.659 35.762,31.349 35.762,30.32C35.762,29.291 36.751,27.719 36.751,27.719C36.751,27.719 37.74,29.293 37.74,30.32Z"
+      android:strokeAlpha="0.6"
+      android:fillColor="#ffffff"
+      android:fillAlpha="0.6"/>
+  <path
+      android:pathData="M21.994,33.477C22.406,31.56 24.451,30.106 26.91,30.106C29.322,30.106 31.335,31.507 31.8,33.367"
+      android:strokeWidth="1.60396"
+      android:fillColor="#00000000"
+      android:strokeColor="#1A211B"
+      android:strokeLineCap="round"/>
+</vector>
