+++ app/src/main/res/drawable/ic_emoji_mint_what_fiter_54.xml
@@ -1,0 +1,99 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="54dp"
+    android:height="54dp"
+    android:viewportWidth="54"
+    android:viewportHeight="54">
+  <group>
+    <clip-path
+        android:pathData="M0,0h54v54h-54z"/>
+    <path
+        android:pathData="M53.343,31.734C57.166,44.785 45.048,56.903 31.997,53.08C28.84,52.155 25.483,52.155 22.326,53.08C9.275,56.903 -2.843,44.785 0.98,31.734C1.905,28.577 1.905,25.219 0.98,22.062C-2.843,9.011 9.275,-3.107 22.326,0.716C25.483,1.641 28.84,1.641 31.997,0.716C45.048,-3.107 57.166,9.011 53.343,22.062C52.419,25.219 52.419,28.577 53.343,31.734Z">
+      <aapt:attr name="android:fillColor">
+        <gradient 
+            android:startX="5.669"
+            android:startY="5.405"
+            android:endX="52.327"
+            android:endY="52.063"
+            android:type="linear">
+          <item android:offset="0" android:color="#FFA0F2B7"/>
+          <item android:offset="0.058" android:color="#FF99EFAF"/>
+          <item android:offset="0.225" android:color="#FF8AE89D"/>
+          <item android:offset="0.346" android:color="#FF85E597"/>
+          <item android:offset="0.36" android:color="#FF83E395"/>
+          <item android:offset="0.535" android:color="#FF6ED486"/>
+          <item android:offset="0.651" android:color="#FF67CE80"/>
+          <item android:offset="1" android:color="#FF5FB772"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <path
+        android:pathData="M36.917,0.002V1.539C41.766,1.539 46.436,3.908 49.406,7.875C52.374,11.834 53.269,16.849 51.868,21.631C50.858,25.077 50.858,28.721 51.868,32.167C53.269,36.949 52.371,41.964 49.406,45.923C46.434,49.89 41.766,52.259 36.917,52.259C35.422,52.259 33.912,52.04 32.431,51.604C30.72,51.102 28.948,50.847 27.165,50.847C25.382,50.847 23.609,51.102 21.896,51.604C20.415,52.037 18.904,52.259 17.41,52.259C12.561,52.259 7.89,49.89 4.92,45.923C1.953,41.964 1.058,36.949 2.458,32.167C3.469,28.721 3.469,25.077 2.458,21.631C1.058,16.849 1.956,11.834 4.92,7.875C7.893,3.908 12.561,1.539 17.41,1.539C18.904,1.539 20.415,1.758 21.896,2.194C23.607,2.697 25.379,2.951 27.162,2.951C28.945,2.951 30.717,2.697 32.428,2.194C33.909,1.761 35.42,1.539 36.914,1.539L36.917,0.002ZM36.917,0.002C35.313,0.002 33.663,0.229 31.998,0.718C30.418,1.181 28.79,1.411 27.162,1.411C25.534,1.411 23.906,1.181 22.326,0.718C20.661,0.232 19.011,0.002 17.407,0.002C6.444,0.002 -2.356,10.679 0.98,22.064C1.905,25.222 1.905,28.579 0.98,31.736C-2.356,43.122 6.441,53.799 17.407,53.799C19.011,53.799 20.661,53.572 22.326,53.082C23.906,52.62 25.534,52.39 27.162,52.39C28.79,52.39 30.418,52.62 31.998,53.082C33.663,53.569 35.313,53.799 36.917,53.799C47.88,53.799 56.68,43.122 53.344,31.736C52.419,28.579 52.419,25.222 53.344,22.064C56.68,10.676 47.882,0.002 36.917,0.002Z"
+        android:strokeAlpha="0.8"
+        android:fillAlpha="0.8">
+      <aapt:attr name="android:fillColor">
+        <gradient 
+            android:startX="27.163"
+            android:startY="-5.985"
+            android:endX="27.163"
+            android:endY="63.058"
+            android:type="linear">
+          <item android:offset="0" android:color="#FFFFFFFF"/>
+          <item android:offset="0.243" android:color="#FFFDFDFD"/>
+          <item android:offset="0.386" android:color="#FFF4F8F5"/>
+          <item android:offset="0.504" android:color="#FFE6EEE8"/>
+          <item android:offset="0.608" android:color="#FFD2E0D5"/>
+          <item android:offset="0.703" android:color="#FFB8CEBC"/>
+          <item android:offset="0.79" android:color="#FF98B89E"/>
+          <item android:offset="0.873" android:color="#FF719E7B"/>
+          <item android:offset="0.95" android:color="#FF468152"/>
+          <item android:offset="1" android:color="#FF266B34"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <path
+        android:pathData="M27.388,38.426C30.02,38.426 32.154,35.808 32.154,32.579C32.154,29.351 30.02,26.733 27.388,26.733C24.756,26.733 22.622,29.351 22.622,32.579C22.622,35.808 24.756,38.426 27.388,38.426Z">
+      <aapt:attr name="android:fillColor">
+        <gradient 
+            android:centerX="27.388"
+            android:centerY="32.579"
+            android:gradientRadius="6.718"
+            android:type="radial">
+          <item android:offset="0" android:color="#FF1F4925"/>
+          <item android:offset="1" android:color="#FF143119"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <path
+        android:pathData="M27.388,38.452C30.02,38.452 32.154,35.834 32.154,32.605C32.154,29.376 30.02,26.759 27.388,26.759C24.756,26.759 22.622,29.376 22.622,32.605C22.622,35.834 24.756,38.452 27.388,38.452Z"
+        android:fillAlpha="0.8">
+      <aapt:attr name="android:fillColor">
+        <gradient 
+            android:startX="27.387"
+            android:startY="26.979"
+            android:endX="27.387"
+            android:endY="39.752"
+            android:type="linear">
+          <item android:offset="0" android:color="#FF143018"/>
+          <item android:offset="1" android:color="#FF468751"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <path
+        android:pathData="M35.403,24.96C36.745,24.96 37.833,23.872 37.833,22.53C37.833,21.188 36.745,20.1 35.403,20.1C34.061,20.1 32.973,21.188 32.973,22.53C32.973,23.872 34.061,24.96 35.403,24.96Z"
+        android:fillColor="#1A211B"/>
+    <path
+        android:pathData="M18.929,24.96C20.271,24.96 21.359,23.872 21.359,22.53C21.359,21.188 20.271,20.1 18.929,20.1C17.587,20.1 16.499,21.188 16.499,22.53C16.499,23.872 17.587,24.96 18.929,24.96Z"
+        android:fillColor="#1A211B"/>
+    <path
+        android:pathData="M34.579,22.321C34.951,22.321 35.253,22.02 35.253,21.648C35.253,21.276 34.951,20.974 34.579,20.974C34.207,20.974 33.905,21.276 33.905,21.648C33.905,22.02 34.207,22.321 34.579,22.321Z"
+        android:strokeAlpha="0.5"
+        android:fillColor="#5FB772"
+        android:fillAlpha="0.5"/>
+    <path
+        android:pathData="M18.115,22.321C18.486,22.321 18.788,22.02 18.788,21.648C18.788,21.276 18.486,20.974 18.115,20.974C17.742,20.974 17.441,21.276 17.441,21.648C17.441,22.02 17.742,22.321 18.115,22.321Z"
+        android:strokeAlpha="0.5"
+        android:fillColor="#5FB772"
+        android:fillAlpha="0.5"/>
+  </group>
+</vector>
